Completed all exercises in Part 2-a. Moved Course component and relevant subcomponents to a separate module (subcomponents are in the same file as Course).
